+++ src/ExpoYandexMapKitView.web.tsx
@@ -23,6 +23,7 @@ export const YandexMapView = React.forwardRef<YandexMapViewRef, YandexMapViewPro
         takeSnapshot: unsupported,
         selectGeoObject: unsupported,
         deselectGeoObject: unsupported,
+        fitToCoordinates: unsupported,
       };
     }, []);
     return null;